:recycle: 以下のリファクタリングを実施、詳細は下記の通り
- TemplateComponentに処理を移譲するように変更
- Template内では index.ts で必要なComponentだけをexportするように変更
- OrgUploadTemplateにねこ画像をJSXに渡すように変更
