+++ src/templates/TermsOrPrivacyTemplate/index.ts
@@ -1,0 +1,1 @@
+export { TermsOrPrivacyTemplate } from './TermsOrPrivacyTemplate';
