+++ src/templates/UploadTemplate/index.ts
@@ -1,0 +1,1 @@
+export { UploadTemplate } from './UploadTemplate';
